Rename and move the kryo serializer to the correct location.
